health_wellness_agent/requirements.txt
# --- Simplified requirements ---
streamlit==1.34.0
pydantic>=2.0.0
python-dotenv==1.0.1
asyncio-mqtt>=0.13.0
typing-extensions>=4.0.0
pandas>=2.0.0
plotly>=5.0.0
fpdf==1.7.2
google-generativeai==0.4.1
groq>=0.8.0


